# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user": {
      "get": {
        "operationId": "UsersController_findOne",
        "summary": "Get current user info",
        "description": "The logged in user info.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserDto"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "summary": "Change current user info",
        "description": "**Remember that if you change the `role` field you will have to login again.**\n\n```json\n{\"role\": \"MARKETING\"}\n```",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "The fields to change",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The changed user info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserDto"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/user/all": {
      "get": {
        "operationId": "UsersController_findAll",
        "summary": "Get all users",
        "parameters": [
          {
            "name": "fromId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "count",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CreateUserDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/user/{username}": {
      "get": {
        "operationId": "UsersController_getUser",
        "summary": "Information about a user via username",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserDto"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/user/{username}/online": {
      "get": {
        "operationId": "UsersController_checkOnline",
        "summary": "Check if a user is online via username",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "username": "string",
                  "online": true
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
...
    "/requests": {
      "post": {
        "operationId": "RequestsController_createRequest",
        "summary": "Create a request to management",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffRequest"
                }
              }
            }
          }
        },
        "tags": [
          "requests"
        ]
      },
      "get": {
        "operationId": "RequestsController_getAll",
        "summary": "Get all requests",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StaffRequest"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "requests"
        ]
      }
    },
    "/requests/{id}": {
      "get": {
        "operationId": "RequestsController_getOne",
        "summary": "Get one requests",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffRequest"
                }
              }
            }
          }
        },
        "tags": [
          "requests"
        ]
      },
      "delete": {
        "operationId": "RequestsController_delete",
        "summary": "Delete a request",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffRequest"
                }
              }
            }
          }
        },
        "tags": [
          "requests"
        ]
      }
    },
    "/requests/attachment/{attachmentId}": {
      "get": {
        "operationId": "RequestsController_getAttachment",
        "summary": "Get attachment information",
        "parameters": [
          {
            "name": "attachmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "name": "name",
                  "filename": "filename",
                  "text": "content"
                }
              }
            }
          }
        },
        "tags": [
          "requests"
        ]
      }
    },
...
    "/admin/requests/{id}": {
      "get": {
        "operationId": "AdminController_getRequest",
        "summary": "Get a specific request",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffRequest"
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
...
    "/admin/user/{id}": {
      "delete": {
        "operationId": "AdminController_deleteUser",
        "summary": "Delete User Account",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Attendance API",
    "description": "API documentation for attendance backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "MANAGEMENT",
              "MARKETING",
              "DEVELOPMENT",
              "SALES",
              "RND"
            ]
          }
        },
        "required": [
          "name",
          "username",
          "email",
          "password",
          "role"
        ]
      },
...
      "CreateRequestDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "content",
          "filename"
        ]
      },
      "StaffRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "userId": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "APPROVED",
              "PENDING",
              "REJECTED",
              "CONSULTATION_REQUIRED"
            ]
          }
        },
        "required": [
          "title",
          "content",
          "filename",
          "userId",
          "status"
        ]
      },
...
    }
  }
}